Python: Replace publish action (#268)
Replaces `cucumber/action-publish-pypi` with
`pypa/gh-action-pypi-publish@release/v1`. The motivation for using
actions in the cucumber org is to ensure that we do not hand release
tokens to untrusted code. As the party publishing our python packages,
the Python Package Authority can be trusted. Additionally, their action
uses trusted publishers which authorizes GitHub with OIDC so no
long-lived tokens are used.